Factor out the setup code from __main__.py.
As we plan to have additional daemons, we should factor out the shared
code from __main__.py to be reused elsewhere. Namely, we can reuse the
argument parsing, logger initialization, and loading the
configuration.
